initial implementation of leaderboard

+++ src/data/Result.java
@@ -3,9 +3,9 @@ package data;
 public class Result {
 
     private Planner planner;
-    private double result;
+    private int result;
 
-    public Result(Planner planner, double result) {
+    public Result(Planner planner, int result) {
         this.planner = planner;
         this.result = result;
     }
@@ -14,7 +14,7 @@ public class Result {
         return planner;
     }
 
-    public double getResult() {
+    public int getResult() {
         return result;
     }
 }
